requirements.txt
streamlit
Pillow
easyocr